Created/ implemented play function
